refactor: stateStore의 위치를 domain 내부로 이동
