+++ src/main.js
@@ -17,7 +17,7 @@ import {
 } from './constants.js';
 import { RESTAURANTS } from './database/restaurantData.js';
 import eventHandlers from './eventHandlers/eventHandlers.js';
-import stateStore from './stateStore.js';
+import stateStore from './domain/stateStore.js';
 import {
   createButton,
   createHeader,
